pyproject.toml
[project]
name = "youtube-title-optimizer"
version = "1.0.0"
description = "AI-powered YouTube title optimizer using LangGraph and Claude"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.1.0",
    "jinja2>=3.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "isort>=5.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
